+++ Project/src/main/java/mx/LemonTrees/Project/Repository/Pago_TrabajadorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Pago_TrabajadorRepository extends CrudRepository<Pago_Trabajador,Long> {
+public interface Pago_TrabajadorRepository extends CrudRepository<Pago_Trabajador,Integer> {
     
 }
